backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.118.0"}
psycopg = {extras = ["binary"], version = "^3.2.10"}
pdfplumber = "^0.11.7"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pillow = "^11.3.0"
openai = "^2.1.0"
anthropic = "^0.69.0"
pandas = "^2.3.3"
matplotlib = "^3.10.6"
openpyxl = "^3.1.5"
python-multipart = "^0.0.20"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
